Fix initial load of rotated image
When user have crop data which includes rotate degree and user try to create new Cropper with that data the image does not show up fully. This happens when rotate degree is 90 or 270. In initCanvas "this.image.rotate" will always be undefined on the first load. Code change: use rotate from image or use it from options that user passed in and update aspectRatio correctly.